Convert to use compose watch (#15)
* Adjust col widths for the item display to provide enough space

* Fix cache location for yarn installs

* Update the project to use Compose Watch

This change is due to a limitation of the WSL integration of file
mounts. Essentially, changes made to files on the host when update
files in the container, but the lack of filesystem event
propagation causes the vite/nodemon file watchers to not observe
the change and perform rebuilds/restarts.

Instead of reverting to file polling (which would have a major CPU
impact), we decided to use Compose Watch for a "define once, run
anywhere" experience.